noUnzip on RunWorkflow

+++ src/RunWorkflow/RunWorkflow.js
@@ -31,7 +31,7 @@ async function main()
 {
     try
     {   
-        let ghToken, workflowName, ref, timeout, dlArtifacts, workflowInputs;
+        let ghToken, workflowName, ref, timeout, dlArtifacts, workflowInputs, noUnzip;
         if(tl.getVariable("Agent.Version")) //Running from the agent
         {
             const auth = tl.getEndpointAuthorization(tl.getInput("gh"), false).parameters;
@@ -50,6 +50,7 @@ async function main()
                 timeout = 120;
             dlArtifacts = tl.getBoolInput("dlArtifacts");
             workflowInputs = tl.getInput("inputs");
+            noUnzip = tl.getBoolInput("noUnzip");
         }
         else //Interactive run
         {
@@ -60,6 +61,7 @@ async function main()
             timeout = 120;
             dlArtifacts = true;
             workflowInputs = "";
+            noUnzip = true;
         }
 
         if(workflowInputs)
@@ -181,20 +183,27 @@ async function main()
                     await new Promise((resolve, err) => {fst.on("finish", resolve); fst.on("error", err); zipDownload.data.on("error", err)});
 
                     const st = fs.statSync(zipFileName);
-                    console.log(`Downloaded artifact ${artName} - ${st.size} bytes, unzipping...`);
-
-                    //Unzip it. Ideally, unzip straight from the web response stream would be cute.
-                    if(fs.existsSync(artName))
+                    if(noUnzip)
                     {
-                        if(fs.statSync(artName).isDirectory())
-                            rimraf.sync(artName);
-                        else
-                            fs.unlinkSync(artName);
+                        console.log(`Downloaded artifact ${artName} - ${st.size} bytes.`);
                     }
-                    fs.mkdirSync(artName);
-                    const zipFile = new nodeStreamZip.async({file: zipFileName});
-                    await zipFile.extract(null, artName);
-                    fs.unlinkSync(zipFileName);
+                    else
+                    {
+                        console.log(`Downloaded artifact ${artName} - ${st.size} bytes, unzipping...`);
+
+                        //Unzip it. Ideally, unzip straight from the web response stream would be cute.
+                        if(fs.existsSync(artName))
+                        {
+                            if(fs.statSync(artName).isDirectory())
+                                rimraf.sync(artName);
+                            else
+                                fs.unlinkSync(artName);
+                        }
+                        fs.mkdirSync(artName);
+                        const zipFile = new nodeStreamZip.async({file: zipFileName});
+                        await zipFile.extract(null, artName);
+                        fs.unlinkSync(zipFileName);
+                    }
                 }
                 catch(exc)
                 {
